+++ src/com/maddyhome/idea/vim/extension/VimExtensionRegistrar.kt
@@ -22,14 +22,16 @@ import com.intellij.openapi.extensions.ExtensionPointListener
 import com.intellij.openapi.extensions.PluginDescriptor
 import com.maddyhome.idea.vim.VimPlugin
 import com.maddyhome.idea.vim.key.MappingOwner.Plugin.Companion.remove
+import com.maddyhome.idea.vim.option.Option
+import com.maddyhome.idea.vim.option.OptionsManager
 import com.maddyhome.idea.vim.option.OptionsManager.addOption
 import com.maddyhome.idea.vim.option.OptionsManager.isSet
 import com.maddyhome.idea.vim.option.OptionsManager.removeOption
 import com.maddyhome.idea.vim.option.ToggleOption
-import java.util.*
 
 object VimExtensionRegistrar {
   private val registeredExtensions: MutableSet<String> = HashSet()
+  private val extensionAliases = HashMap<String, String>()
   private var extensionRegistered = false
   private val logger = logger<VimExtensionRegistrar>()
 
@@ -56,6 +58,7 @@ object VimExtensionRegistrar {
     if (name in registeredExtensions) return
 
     registeredExtensions.add(name)
+    registerAliases(name)
     val option = ToggleOption(name, name, false)
     option.addOptionChangeListener { _, _ ->
       for (extensionInListener in VimExtension.EP_NAME.extensionList) {
@@ -76,9 +79,24 @@ object VimExtensionRegistrar {
     val name = extension.name
     if (name !in registeredExtensions) return
     registeredExtensions.remove(name)
+    removeAliases(extension)
     extension.dispose()
     removeOption(name)
     remove(name)
     logger.info("IdeaVim extension '$name' disposed")
+  }
+
+  fun getToggleByAlias(alias: String): ToggleOption? {
+    val name = extensionAliases[alias] ?: return null
+    return OptionsManager.getOption(name) as ToggleOption?
+  }
+
+  private fun registerAliases(name: String) {
+    val extension = VimExtension.EP_NAME.findFirstSafe { it.name == name } ?: return
+    extension.aliases.forEach { alias -> extensionAliases[alias] = name }
+  }
+
+  private fun removeAliases(extension: VimExtension) {
+    extension.aliases.forEach { extensionAliases.remove(it) }
   }
 }