+++ src/com/maddyhome/idea/vim/extension/highlightedyank/VimHighlightedYank.kt
@@ -83,6 +83,12 @@ class VimHighlightedYank: VimExtension, VimYankListener, VimInsertListener {
 
   override fun getName() = "highlightedyank"
 
+  override fun getAliases(): Set<String> = setOf(
+    "https://github.com/machakann/vim-highlightedyank",
+    "machakann/vim-highlightedyank",
+    "vim-highlightedyank"
+  )
+
   override fun init() {
     VimPlugin.getYank().addListener(this)
     VimPlugin.getChange().addInsertListener(this)