+++ src/com/maddyhome/idea/vim/extension/multiplecursors/VimMultipleCursorsExtension.kt
@@ -59,6 +59,14 @@ private const val ALL_OCCURRENCES = "<Plug>AllOccurrences"
 class VimMultipleCursorsExtension : VimExtension {
   override fun getName() = "multiple-cursors"
 
+  override fun getAliases(): Set<String> {
+    return setOf(
+      "https://github.com/terryma/vim-multiple-cursors",
+      "terryma/vim-multiple-cursors",
+      "vim-multiple-cursors"
+    )
+  }
+
   override fun init() {
     putExtensionHandlerMapping(MappingMode.NXO, parseKeys(NEXT_WHOLE_OCCURRENCE), owner, NextOccurrenceHandler(), false)
     putExtensionHandlerMapping(MappingMode.NXO, parseKeys(NEXT_OCCURRENCE), owner, NextOccurrenceHandler(whole = false), false)