+++ src/com/maddyhome/idea/vim/extension/textobjentire/VimTextObjEntireExtension.java
@@ -21,6 +21,7 @@ package com.maddyhome.idea.vim.extension.textobjentire;
 import com.intellij.openapi.actionSystem.DataContext;
 import com.intellij.openapi.editor.Caret;
 import com.intellij.openapi.editor.Editor;
+import com.intellij.util.containers.ContainerUtil;
 import com.maddyhome.idea.vim.command.*;
 import com.maddyhome.idea.vim.common.TextRange;
 import com.maddyhome.idea.vim.extension.VimExtension;
@@ -33,6 +34,7 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.EnumSet;
+import java.util.Set;
 
 import static com.maddyhome.idea.vim.extension.VimExtensionFacade.putExtensionHandlerMapping;
 import static com.maddyhome.idea.vim.extension.VimExtensionFacade.putKeyMapping;
@@ -65,6 +67,16 @@ public class VimTextObjEntireExtension implements VimExtension {
   public @NotNull
   String getName() {
     return "textobj-entire";
+  }
+
+  @Override
+  public Set<String> getAliases() {
+    return ContainerUtil.set(
+      "https://github.com/kana/vim-textobj-entire",
+      "kana/vim-textobj-entire",
+      "vim-textobj-entire",
+      "https://www.vim.org/scripts/script.php?script_id=2610"
+    );
   }
 
   @Override